+++ facebook/__init__.py
@@ -78,13 +78,14 @@ class GraphAPI(object):
     """
 
     def __init__(self, access_token=None, timeout=None, version=None,
-                 proxies=None):
+                 proxies=None, session=None):
         # The default version is only used if the version kwarg does not exist.
         default_version = VALID_API_VERSIONS[0]
 
         self.access_token = access_token
         self.timeout = timeout
         self.proxies = proxies
+        self.session = session or requests.Session()
 
         if version:
             version_regex = re.compile("^\d\.\d$")
@@ -197,7 +198,7 @@ class GraphAPI(object):
         """Fetches the current version number of the Graph API being used."""
         args = {"access_token": self.access_token}
         try:
-            response = requests.request(
+            response = self.session.request(
                 "GET",
                 FACEBOOK_GRAPH_URL + self.version + "/me",
                 params=args,
@@ -223,7 +224,6 @@ class GraphAPI(object):
         arguments.
 
         """
-        args = args or {}
 
         if post_args is not None:
             method = "POST"
@@ -239,13 +239,14 @@ class GraphAPI(object):
                 args["access_token"] = self.access_token
 
         try:
-            response = requests.request(method or "GET",
-                                        FACEBOOK_GRAPH_URL + path,
-                                        timeout=self.timeout,
-                                        params=args,
-                                        data=post_args,
-                                        proxies=self.proxies,
-                                        files=files)
+            response = self.session.request(
+                method or "GET",
+                FACEBOOK_GRAPH_URL + path,
+                timeout=self.timeout,
+                params=args,
+                data=post_args,
+                proxies=self.proxies,
+                files=files)
         except requests.HTTPError as e:
             response = json.loads(e.read())
             raise GraphAPIError(response)