+++ docs/changes.rst
@@ -8,6 +8,7 @@ Version 3.0.0 (unreleased)
  - Add support for Graph API version 2.8.
  - Remove support for Graph API version 2.1.
  - Change default Graph API version to 2.2.
+ - Add support for requests' sessions.
 
 Version 2.0.0 (2016-08-08)
 ==============================